+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 
 # CHANGELOG
 
+## v3.7.0
+
+* Add `SearchFlip::Connection#bulk` to allow more clean bulk indexing to
+  multiple indices at once
+
 ## v3.6.0
 
 * Support Elasticsearch v8